Fix errors wrapping when applying resources (#2)
Errors are properly wrapped now, so it is possible to check with `errors.Is`
or `errors.As` if there is a given error in the list of returned errors. And it is
also possible now to `Unwrap() []error` to get the actual list of errors.
Apply is not interrupted now if Get or NeedsUpdate fail for a resource. A
result with unknown action and an error is returned for these resources.

Go 1.19 is not supported anymore.