[breaking] Use standard contexts
Replace the concept of context from the library, and use standard
contexts instead.

The extended functionality provided before by these contexts is provided
now by the new concept of "Runtime".

A new helper is added to the manager to enrich a standard context with a
runtime. Also a helper function is added to get a runtime from a
context.

This change is breaking for implementation of resource types. It should
be transparent for users of resources.